[bugfix] fix a naming error of header right area(related to login) - problem: component naming error (related to login & user info) - modification: define header riath area component -- change the position of RightArea div styled def -- change naming of RightAreaComponent, RightArea
