+++ src/components/organisms/main/Header.tsx
@@ -2,23 +2,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
 import Logo from '../../../assets/logo.svg';
-import Button from '../../atoms/Button';
-import { useAppSelector } from '../../../redux/hooks';
-
-const RigthAreaComponent = () => {
-  const isCertified = useAppSelector(state => state.userState.is_certified);
-  return !isCertified ? (
-    <Link to="/login">
-      <Button label="로그인" />
-    </Link>
-  ) : (
-    <>
-      <button>새 글 작성</button>
-      <button>User Icon</button>
-      <button>User Menu</button>
-    </>
-  );
-};
+import HeaderRightArea from './HeaderRightArea';
 
 const Header = () => {
   return (
@@ -27,10 +11,7 @@ const Header = () => {
         <Link to="/">
           <img src={Logo} alt="logo" />
         </Link>
-        <RightArea>
-          <button>찾기</button>
-          <RigthAreaComponent />
-        </RightArea>
+        <HeaderRightArea />
       </InnerBox>
     </HeaderStyle>
   );
@@ -47,9 +28,4 @@ const InnerBox = styled.div`
   display: flex;
   align-items: center;
   justify-content: space-between;
-`;
-
-const RightArea = styled.div`
-  display: flex;
-  align-items: center;
 `;
